Rename the cccl.c.parallel c2h.h header to avoid confusion.
